Move config classes to org.joshsim.engine.config package
Reorganizes configuration-related classes for better structure by:
- Moving NoOpConfigGetter from org.joshsim.lang.bridge to org.joshsim.engine.config
- Moving JshcConfigGetter from org.joshsim.precompute to org.joshsim.engine.config
- Updating all import statements in affected files
- Moving test files to match new package structure

This consolidates all config functionality in one package while maintaining
the ConfigGetter interface in the bridge abstraction layer.
