+++ src/main/java/org/joshsim/JoshSimFacadeUtil.java
@@ -19,7 +19,7 @@ import org.joshsim.lang.io.CombinedExportFacade;
 import org.joshsim.lang.io.InputOutputLayer;
 import org.joshsim.lang.parse.JoshParser;
 import org.joshsim.lang.parse.ParseResult;
-import org.joshsim.precompute.JshcConfigGetter;
+import org.joshsim.engine.config.JshcConfigGetter;
 import org.joshsim.precompute.JshdExternalGetter;
