+++ src/main/java/org/joshsim/command/PreprocessCommand.java
@@ -42,7 +42,7 @@ import org.joshsim.precompute.BinaryGridSerializationStrategy;
 import org.joshsim.precompute.DataGridLayer;
 import org.joshsim.precompute.ExtentsTransformer;
 import org.joshsim.precompute.GridCombiner;
-import org.joshsim.precompute.JshcConfigGetter;
+import org.joshsim.engine.config.JshcConfigGetter;
 import org.joshsim.precompute.JshdExternalGetter;
 import org.joshsim.precompute.PatchKeyConverter;
 import org.joshsim.precompute.StreamToPrecomputedGridUtil;
