+++ src/test/java/org/joshsim/engine/config/JshcConfigGetterTest.java
@@ -4,7 +4,7 @@
  * @license BSD-3-Clause
  */
 
-package org.joshsim.precompute;
+package org.joshsim.engine.config;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
